Accommodate moving outbox to toolkit
